+++ DestroyusProWhiteList.lua
@@ -1,0 +1,5 @@
+function _DestroyusProItems.InitWhiteList()
+    _G["DestroyusProWhiteList"] = {}
+    _G["DestroyusProWhiteList"]["activated"] = false
+    _G["DestroyusProWhiteList"]["itemList"] = {}
+end