+++ CHANGELOG.md
@@ -5,14 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 1.0.0 - 10.05.2020
+## New version 1.1.0 - 31.08.2020 Shadowlands
 
-### Added
+### Added in 1.1.0
+
+- Added option to select and manage whitelist for disenchanting
+
+### Notes
+
+- When whitelist is active disenchanting will be performed only on items from this list
+- Whitelist is activated per character
+- Blacklist is activated in global scope
+
+## New version 1.0.0 - 10.05.2020
+
+### Added in 1.0.0
 
 - Added option to manage blacklist for disenchanting by adding or removing itemID
 - Added check to skip items saved as set from Equipment Manager
 
-### Changed
+### Changed in 1.0.0
 
 - Updated function for detecting items which can be disenchanted
 - Updated addon frame